pyproject.toml
[project]
name = "mcp-powerpoint"
version = "1.0.0"
description = "MCP server for PowerPoint file manipulation and shape naming using Ollama LLM"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    {name = "MCP PowerPoint Tools"}
]
keywords = ["mcp", "powerpoint", "pptx", "json", "llm", "ollama", "shape-naming"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Office Suites",
    "Topic :: Text Processing :: Markup"
]
dependencies = [
    "fastmcp>=2.12.2",
    "ollama>=0.5.3",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "python-pptx>=1.0.2",
    "torch>=2.8.0",
    "transformers>=4.56.1",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.7.0",
    "nltk>=3.9.1",
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "aiofiles>=23.2.1",
    "jinja2>=3.1.3",
    "python-multipart>=0.0.9",
]

[tool.setuptools.packages.find]
include = ["mcp_powerpoint*"]

[project.scripts]
mcp-powerpoint = "mcp_powerpoint.__main__:main"

[project.urls]
Homepage = "https://github.com/your-username/mcp-powerpoint"
Repository = "https://github.com/your-username/mcp-powerpoint.git"
Issues = "https://github.com/your-username/mcp-powerpoint/issues"

[tool.mcp.servers.powerpoint]
command = "uvx"
args = ["mcp-powerpoint", "--server", "powerpoint"]
description = "PowerPoint file conversion and manipulation tools"

[tool.mcp.servers.shape-naming]
command = "uvx"
args = ["mcp-powerpoint", "--server", "shape-naming"]
description = "Shape naming tools using Ollama LLM"

[tool.uv.workspace]
members = [
    ".",
]

[tool.uv.sources]
mcp-powerpoint = { workspace = true }

[dependency-groups]
dev = [
    "mcp-powerpoint",
]
